fixes issue #225: first attempt to eliminate default exports
