+++ src/helpers/attachto.ts
@@ -46,4 +46,3 @@ export function attachTo(target: Element, vnode: VNode): VNode {
   hook.destroy = destroy;
   return vnode;
 };
-export default attachTo;
