+++ src/modules/class.ts
@@ -25,4 +25,3 @@ function updateClass(oldVnode: VNode, vnode: VNode): void {
 }
 
 export const classModule = {create: updateClass, update: updateClass} as Module;
-export default classModule;
