+++ src/modules/dataset.ts
@@ -36,4 +36,3 @@ function updateDataset(oldVnode: VNode, vnode: VNode): void {
 }
 
 export const datasetModule = {create: updateDataset, update: updateDataset} as Module;
-export default datasetModule;
