+++ src/modules/props.ts
@@ -26,4 +26,3 @@ function updateProps(oldVnode: VNode, vnode: VNode): void {
 }
 
 export const propsModule = {create: updateProps, update: updateProps} as Module;
-export default propsModule;
