+++ src/snabbdom.bundle.ts
@@ -1,16 +1,15 @@
 import {init} from './snabbdom';
-import {attributesModule} from './modules/attributes'; // for setting attributes on DOM elements
-import {classModule} from './modules/class'; // makes it easy to toggle classes
-import {propsModule} from './modules/props'; // for setting properties on DOM elements
-import {styleModule} from './modules/style'; // handles styling on elements with support for animations
-import {eventListenersModule} from './modules/eventlisteners'; // attaches event listeners
-import {h} from './h'; // helper function for creating vnodes
+import {attrs} from './snabbdom'; // for setting attributes on DOM elements
+import {classes} from './snabbdom'; // makes it easy to toggle classes
+import {props} from './snabbdom'; // for setting properties on DOM elements
+import {styles} from './snabbdom'; // handles styling on elements with support for animations
+import {events} from './snabbdom'; // attaches event listeners
+import {h} from './snabbdom'; // helper function for creating vnodes
 var patch = init([ // Init patch function with choosen modules
-  attributesModule,
-  classModule,
-  propsModule,
-  styleModule,
-  eventListenersModule
+  attrs,
+  classes,
+  props,
+  styles,
+  events
 ]) as (oldVNode: any, vnode: any) => any;
 export const snabbdomBundle = { patch, h: h as any };
-export default snabbdomBundle;
